# Conflicts:
#	app/src/main/java/ru/krivorukie/foodcountpp/ActivityPage3.java

+++ app/src/main/java/ru/krivorukie/foodcountpp/UserJSONSerializer.java
@@ -17,7 +17,7 @@ public class UserJSONSerializer {
     };
 
     public static class Product{
-            private double productID;
+            private Integer productID;
             private String name;
             private String size;
             private double energy_value;
@@ -26,7 +26,7 @@ public class UserJSONSerializer {
             private double carbohydrates;
             private double price;
 
-            public double getProductID() {
+            public Integer getProductID() {
                 return productID;
             }
